feat: basic page structure and form for sign up page and moves authentication pages back to page component directory
Relates #61
