+++ src/components/page-components/SignUpPage/SignUpPage.css
@@ -1,5 +1,9 @@
 .sign-up-page {
-  background-color: white;
+  width: 100dvw;
+  height: 100dvh;
+  background-color: var(--base);
+  animation: loginSlideLeft 0.5s ease-out;
+  padding: 20px;
   animation: signUpSlideRight 0.5s ease-out;
 }
